refactor(components/molecule/drawer): api tab
